requirements.txt
streamlit>=1.28.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
Pillow>=10.0.0
pandas>=2.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
openpyxl>=3.1.2
